pyproject.toml
[project]
name = "ollama-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.8.0",
    "openai>=2.11.0",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
]
